Charan (#3)
* Updated Code

* 23/02/2024

* 28/02/2024

+++ src/main/java/com/example/JobSupportBackend/repo/CertificationRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.example.JobSupportBackend.entity.Certification;
 
 @Repository
-public interface CertificationRepository extends JpaRepository<Certification, Long>{
+public interface CertificationRepository extends JpaRepository<Certification, Integer>{
 
 	List<Certification> findByUserEmail(String email);
